+++ CHANGELOG.rst
@@ -92,6 +92,7 @@ Fixed
 - Bug with the 'measure_options' arguments of MLAlgoAssessor and MLAlgoSelection.
 - The constraints names are now correctly formed with the minus sign and offset value if any.
 - The image comparison tests in post/dataset no longer leave the generated files when completed.
+- DesignSpace no longer logs an erroneous warning when unnormalizing an unbounded variable.
 
 Version 3.1.0 (July 2021)
 *************************